changes the extraction.process module name to extraction.loader, which sounds more appropriate
