+++ tests/test_ke_data.py
@@ -3,16 +3,16 @@ import pytest
 from transformerkp.data.extraction.args import KEDataArguments
 from transformerkp.data.extraction.args import InspecKEDataArguments
 from transformerkp.data.extraction.args import NUSKEDataArguments
-from transformerkp.data.extraction.process import KEDataset
-from transformerkp.data.extraction.process import InspecKEDataset
-from transformerkp.data.extraction.process import NUSKEDataset
-from transformerkp.data.extraction.process import KDDKEDataset
-from transformerkp.data.extraction.process import KPCrowdKEDataset
-from transformerkp.data.extraction.process import SemEval2017KEDataset
-from transformerkp.data.extraction.process import SemEval2010KEDataset
-from transformerkp.data.extraction.process import DUC2001KEDataset
-from transformerkp.data.extraction.process import CSTRKEDataset
-from transformerkp.data.extraction.process import PubMedKEDataset
+from transformerkp.data.extraction.loader import KEDataset
+from transformerkp.data.extraction.loader import InspecKEDataset
+from transformerkp.data.extraction.loader import NUSKEDataset
+from transformerkp.data.extraction.loader import KDDKEDataset
+from transformerkp.data.extraction.loader import KPCrowdKEDataset
+from transformerkp.data.extraction.loader import SemEval2017KEDataset
+from transformerkp.data.extraction.loader import SemEval2010KEDataset
+from transformerkp.data.extraction.loader import DUC2001KEDataset
+from transformerkp.data.extraction.loader import CSTRKEDataset
+from transformerkp.data.extraction.loader import PubMedKEDataset
 
 
 @pytest.fixture
